requirements.txt
python-telegram-bot==20.3
requests==2.31.0
openpyxl==3.1.2
urllib3==2.0.4
python-dotenv




